backend/requirements.txt
absl-py==1.4.0
accelerate==0.18.0
aiobotocore==2.4.2
aiofiles==23.1.0
aiohttp==3.8.4
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
albumentations==0.4.6
altair==4.2.2
antlr4-python3-runtime==4.8
appnope==0.1.3
arq==0.25.0
asttokens==2.2.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==23.1.0
awscli==1.25.60
backcall==0.2.0
blinker==1.6.2
botocore==1.27.59
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.4
coloredlogs==15.0.1
contourpy==1.0.7
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
deprecation==2.1.0
diffusers==0.16.1
docutils==0.16
ecdsa==0.18.0
einops==0.3.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.3
frozenlist==1.3.3
fsspec==2023.4.0
ftfy==6.1.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
grpcio==1.54.0
hiredis==2.2.2
huggingface-hub==0.14.1
humanfriendly==10.0
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.4.0
importlib-metadata==6.6.0
iniconfig==2.0.0
invisible-watermark==0.1.5
ipython==8.13.1
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==4.17.3
kiwisolver==1.4.4
kornia==0.6.0
lazy_loader==0.2
# Editable install with no version control (logo-worker==0.20.0)
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.19.0
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
model-index==0.1.11
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.23.1
oauthlib==3.2.2
omegaconf==2.1.1
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.6.0.66
openmim==0.3.7
ordered-set==4.1.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
pluggy==1.0.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.7
pydeck==0.8.1b0
pyDeprecate==0.3.1
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyre-extensions==0.0.23
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-jose==3.3.0
python-keycloak==2.16.0
pytorch-lightning==1.4.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==5.4.1
redis==4.5.4
regex==2023.3.23
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
rich==13.3.5
rsa==4.7.2
s3transfer==0.6.0
scikit-image==0.20.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
streamlit==1.22.0
sympy==1.11.1
tabulate==0.9.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
test-tube==0.7.5
tifffile==2023.4.12
timm==0.6.13
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torch-fidelity==0.3.0
torchaudio==0.13.1
torchmetrics==0.6.0
torchvision==0.14.1
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
ujson==5.7.0
urllib3==1.26.15
urwid==2.1.2
validators==0.20.0
wcwidth==0.2.6
Werkzeug==2.3.2
wrapt==1.15.0
xformers==0.0.16
yarl==1.9.2
zipp==3.15.0
